マップ描画処理を別クラスに切り出す
