+++ packages/map-editor/src/MapCanvas.ts
@@ -6,6 +6,7 @@ import { Arrangements } from './Brushes/Arrangements/Arrangements'
 import { Brush } from './Brushes/Brush'
 import { Arrangement, isTiledMapDataRequired, isAutoTileRequired, isAutoTilesRequired } from './Brushes/Arrangements/Arrangement'
 import { DefaultArrangement } from './Brushes/Arrangements/DefaultArrangement'
+import { MapRenderer } from './MapRenderer'
 
 export class MapCanvas {
   private _ctx = this.canvas.getContext('2d') as CanvasRenderingContext2D
@@ -13,15 +14,16 @@ export class MapCanvas {
   private _isMouseDown = false
   private _brush: Brush = new Pen()
   private _arrangement: Arrangement = new DefaultArrangement()
-  private _backgroundRgba = {r: 255, g: 255, b: 255, a: 1.0}
   private _lastMapChipPosition = {x: -1, y: -1}
+  private _renderer = new MapRenderer(this._project.tiledMap)
 
   constructor(
     private _project: Project,
     private canvas: HTMLCanvasElement,
     private secondaryCanvas: HTMLCanvasElement,
+    private coliderCanvas: HTMLCanvasElement
   ) {
-    this._project.registerRenderAllCallback(() => this.renderAll())
+    this._project.registerRenderAllCallback(() => this._renderer.renderAll(this._ctx))
   }
 
   public setBrushFromName(brushName: string) {
@@ -91,7 +93,7 @@ export class MapCanvas {
     this.clearSecondaryCanvas()
     this._brush.mouseMove(chipPosition.x, chipPosition.y).forEach(paint => {
       const chip = paint.chip
-      this._putOrClearChipToCanvas(this._secondaryCanvasCtx, chip, paint.x, paint.y, true)
+      this._renderer.putOrClearChipToCanvas(this._secondaryCanvasCtx, chip, paint.x, paint.y, true)
     })
 
     this._lastMapChipPosition = chipPosition
@@ -116,96 +118,11 @@ export class MapCanvas {
 
   putChip(mapChip: MapChip | null, chipX: number, chipY: number) {
     this._project.tiledMap.put(mapChip, chipX, chipY)
-    this._putOrClearChipToCanvas(this._ctx, mapChip, chipX, chipY)
-  }
-
-  renderAll() {
-    this._project.tiledMap.data.items.forEach((value, index) => {
-      const position = this._project.tiledMap.data.convertMapNumberToPosition(index)
-      this._putOrClearChipToCanvas(this._ctx, value, position.x, position.y)
-    })
+    this._renderer.putOrClearChipToCanvas(this._ctx, mapChip, chipX, chipY)
   }
 
   private clearSecondaryCanvas() {
     this._secondaryCanvasCtx.clearRect(0, 0, this.secondaryCanvas.width, this.secondaryCanvas.height)
-  }
-
-  private _putOrClearChipToCanvas(ctx: CanvasRenderingContext2D, mapChip: MapChip | null, chipX: number, chipY: number, isTemporaryRendering: boolean = false) {
-    if (mapChip instanceof MapChip) {
-      mapChip.items.forEach(item => {
-        this._putChipToCanvas(ctx, item, chipX, chipY)
-      })
-    } else {
-      this._clearChipToCanvas(ctx, chipX, chipY, isTemporaryRendering)
-    }
-  }
-
-  private _clearChipToCanvas(ctx: CanvasRenderingContext2D, chipX: number, chipY: number, isTemporaryRendering: boolean) {
-    const position = this._project.tiledMap.convertChipPositionToPixel(chipX, chipY)
-
-    ctx.clearRect(position.x, position.y, this._project.tiledMap.chipWidth, this._project.tiledMap.chipHeight)
-
-    if (isTemporaryRendering) {
-      ctx.fillStyle = `rgba(${this._backgroundRgba.r},${this._backgroundRgba.g},${this._backgroundRgba.b},${this._backgroundRgba.a})`
-      ctx.fillRect(position.x, position.y, this._project.tiledMap.chipWidth, this._project.tiledMap.chipHeight)
-    }
-  }
-
-  private _putChipToCanvas(ctx: CanvasRenderingContext2D, mapChip: MapChipFragment, chipX: number, chipY: number) {
-    const mapChips = this._project.tiledMap.mapChipsCollection.findById(mapChip.chipId)
-    const image = mapChips?.image
-    if (!image) return
-
-    const renderingArea = this._getRenderingArea(mapChip)
-    const position = this._project.tiledMap.convertChipPositionToPixel(chipX, chipY)
-    position.x += renderingArea.destOffsetX
-    position.y += renderingArea.destOffsetY
-
-    ctx.clearRect(position.x, position.y, renderingArea.width, renderingArea.height)
-    ctx.drawImage(
-      image,
-      renderingArea.x,
-      renderingArea.y,
-      renderingArea.width,
-      renderingArea.height,
-      position.x,
-      position.y,
-      renderingArea.width,
-      renderingArea.height
-    )
-  }
-
-  private _getRenderingArea(mapChip: MapChipFragment) {
-    const width = this._project.tiledMap.chipWidth
-    const height = this._project.tiledMap.chipHeight
-    const x = mapChip.x * width
-    const y = mapChip.y * height
-
-    if (mapChip.renderingArea === 15) {
-      return {x, y, width, height, destOffsetX: 0, destOffsetY: 0}
-    }
-
-    const halfWidth = Math.round(width / 2)
-    const halfHeight = Math.round(height / 2)
-
-    switch(mapChip.renderingArea) {
-      case 1:
-        return {x, y, width: halfWidth, height: halfHeight, destOffsetX: 0, destOffsetY: 0}
-      case 2:
-        return {x: x + halfWidth, y, width: halfWidth, height: halfHeight, destOffsetX: halfWidth, destOffsetY: 0}
-      case 3:
-        return {x, y, width, height: halfHeight, destOffsetX: 0, destOffsetY: 0}
-      case 4:
-        return {x, y: y + halfHeight, width: halfWidth, height: halfHeight, destOffsetX: 0, destOffsetY: halfHeight}
-      case 5:
-        return {x, y, width: halfWidth, height, destOffsetX: 0, destOffsetY: 0}
-      case 8:
-        return {x: x + halfWidth, y: y + halfHeight, width: halfWidth, height: halfHeight, destOffsetX: halfWidth, destOffsetY: halfHeight}
-      case 10:
-        return {x: x + halfWidth, y, width: halfWidth, height, destOffsetX: halfWidth, destOffsetY: 0}
-      case 12:
-        return {x, y: y + halfHeight, width, height: halfHeight, destOffsetX: 0, destOffsetY: halfHeight}
-    }
   }
 
   public convertFromCursorPositionToChipPosition(x: number, y: number) {
